+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support for FreeTube (Subscriptions and History only)
+- Support for ViewTube (Subscriptions only)
 
 ## [0.3.3] - 2024-01-16